Remove make local-box goal in Makefile
The previous commit just caches the deb files to the local machine which
is good enough for speeding up subsequent runs of `make local-cluster`.
We will revisit local box creation at some other time if ever. For now,
I'm keeping the builder directory since it can still be useful for when
we want to pre-bake images in the cloud.