Authenticate against declared docker registries
This change allows executors the ability to pull and push to declared
registries so that projects don't have to store that information
themselves. The way it is implemented right now is VERY INSECURE and
must be revisited at a later time. Ideally, jobs should not be able to
access these credentials.